+++ sr_port/mur_write_header_extfmt.c
@@ -116,7 +116,7 @@ void mur_write_header_extfmt(jnl_ctl_list *jctl, FILE *fp, char *fname, int recs
 			assert(save_errno);
 			SNPRINTF(errstr, SIZEOF(errstr),
 				"fwrite() : %s : Expected = %lld : Actual = %lld",
-						(stdout == fp) ? "-STDOUT" : fname, extrlen, ret_size);
+						(stdout == fp) ? "-STDOUT" : fname, (long long)extrlen, (long long)ret_size);
 			rts_error_csa(CSA_ARG(NULL) VARLSTCNT(8)
 						ERR_SYSCALL, 5, LEN_AND_STR(errstr), CALLFROM, save_errno);
 		}
