+++ sr_unix/zshow_rctldump.c
@@ -66,7 +66,7 @@ void	zshow_rctldump(zshow_out *output)
 		nbytes = SNPRINTF(buff, SIZEOF(buff), "# of attached processes  : %d", hdr->nattached);
 		DUMP_ONE_LINE(output, buff, nbytes);
 		nbytes = SNPRINTF(buff, SIZEOF(buff), "Relinkctl shared memory  : shmid: %d  shmlen: 0x%llx",
-			hdr->relinkctl_shmid, hdr->relinkctl_shmlen);
+			hdr->relinkctl_shmid, (unsigned long long)hdr->relinkctl_shmlen);
 		DUMP_ONE_LINE(output, buff, nbytes);
 		shm_hdr = GET_RELINK_SHM_HDR(linkctl);
 		for (i = 0, j = 1; i < NUM_RTNOBJ_SHM_INDEX; i++)
